convert from 17 to 13
