+++ apps/modernization-api/src/main/resources/db/changelog/report-execution-changelog.yml
@@ -25,5 +25,5 @@ databaseChangeLog:
             path: db/report/execution/libraries/nbs_sr_02.sql
             splitStatements: false
         - sqlFile:
-            path: db/report/execution/libraries/nbs_sr_17.sql
+            path: db/report/execution/libraries/nbs_sr_13.sql
             splitStatements: false
