+++ apps/modernization-api/src/main/resources/db/report/execution/03_ODSE_Data_Report_Library_Init.sql
@@ -36,9 +36,9 @@ BEGIN
         ('NBSSR00010.SAS', 'SR10: Multi-Year Line Graph of Disease Cases. Report demonstrates, using a multi-year line graph, the total number of yearly Investigation(s) [both Individual and Summary] for a given disease, by State, irrespective of Case Status.'),
         ('NBSSR00011.SAS', 'SR11: Cases of Selected Diseases By Year Over Time. Report demonstrates, in table form, the total number of Investigation(s) [both Individual and Summary] by calculated MMWR Year irrespective of Case Status.'),
         ('NBSSR00012.SAS', 'SR12: Cases of Selected Disease By County By Year. Report demonstrates, in table form, the total number of Investigation(s) [both Individual and Summary] by Year (calculated based on Event Date year) and by County.'),
+        ('NBSSR00013.SAS', 'Number of Investigations by disease by disease and status.'),
         ('NBSSR00015.SAS', 'Bar chart of 5-Year Case Count, By Month, for Selected Diseases'),
         ('NBSSR00016.SAS', 'Number of Cases and Deaths by Disease, Year Count'),
-        ('NBSSR00017.SAS', 'Number of Investigations by disease by disease and status.'),
         ('NBSSR00018.SAS', 'TB Case Verification Report. Computes statistics about TB cases.'),
         ('NBSSR00019.SAS', 'TB Record Count - Summary Report by Report Date. Count of cases by "countability" and month.'),
         ('NBSSR00020.SAS', 'TB Record Count - Summary Report by Count Date. Count of cases by "countability" and month.'),
